fix(utils): rename platform.py to platform_utils.py to avoid stdlib shadowing
The `onyx.utils.platform` module shadowed the Python stdlib `platform` module,
causing `AttributeError: module 'platform' has no attribute 'system'` when
`uuid` (imported by redis) tried to call `platform.system()`. This crashed the
`supervisord_watchdog_celery_beat` process on container startup.

Closes #10975
