+++ backend/onyx/background/celery/tasks/monitoring/tasks.py
@@ -43,8 +43,8 @@ from onyx.db.search_settings import get_active_search_settings_list
 from onyx.redis.redis_pool import get_redis_client
 from onyx.redis.redis_pool import redis_lock_dump
 from onyx.redis.tenant_redis_client import TenantRedisClient
-from onyx.utils.platform import is_running_in_container
-from onyx.utils.platform import is_running_in_kubernetes
+from onyx.utils.platform_utils import is_running_in_container
+from onyx.utils.platform_utils import is_running_in_kubernetes
 from onyx.utils.telemetry import optional_telemetry
 from onyx.utils.telemetry import RecordType
 from shared_configs.configs import MULTI_TENANT
